scripts: update run_nmsis.sh and download_fpga.sh
How to run:
==========

1. Make sure the following actions are done

  a) fpga_runners section in nmsis_dsp.yaml and nmsis_nn.yaml
     are adapted according to your machine, such as fpga and jtag.
  b) fpga hardware and cpu/fpga jtag debugger are connected
  c) python3 Scripts/Configs/fpga/find_usb.py will help you
     to dump usb device information connected to the pc
  d) make sure you are in /path/to/NMSIS/NMSIS

2. Download neccessary fpga bits using Scripts/Configs/fpga/download_fpga.sh

bash Scripts/Configs/fpga/download_fpga.sh

3. Run test suites on fpga using Scripts/Configs/fpga/run_nmsis.sh

   CFGSET=<cfgset> FPGASET=<fpgaset> CONFIG=<config> bash Scripts/Configs/fpga/run_nmsis.sh

   such as:
   a) CFGSET=mini CONFIG=n300 bash Scripts/Configs/fpga/run_nmsis.sh
      This will run on mini test config, and evaluate on n300 bitstream
   b) CFGSET=full CONFIG=n300,ux900,n900 bash Scripts/Configs/fpga/run_nmsis.sh
      This will run on full test config, and evaluate on n300,ux900,n900
      bitstreams

+++ NMSIS/Scripts/Configs/fpga/download_fpga.sh
@@ -1,4 +1,5 @@
 #!/bin/env bash
+FORCE=${FORCE:-0}
 
 SCRIPTDIR=$(dirname $(readlink -f $BASH_SOURCE))
 FPGABITS=bitstreams
@@ -15,6 +16,11 @@ function wget_file() {
     local file=$2
 
     if [ -f $file ] ; then
+        if [ "x$FORCE" == "x0" ] ; then
+            echo "$file already exist, no need to download it!"
+            return
+        fi
+        echo "$file existed, will remove it, before download it"
         rm -f $file
     fi
     echo "Download $link to $file"
@@ -35,10 +41,32 @@ function wget_and_link {
     popd
 }
 
+function wget2dir {
+    local link=$1
+    local file=$2
+    local dir=$3
+
+    pushd $SCRIPTDIR
+    mkdir -p $dir
+    wget_file $link $dir/$file
+    popd
+}
+
+function make_link_latest {
+    local bitver=${1:-202206}
+    pushd $SCRIPTDIR/$FPGABITS
+    rm -f latest
+    echo "Link $bitver to latest"
+    ln -s $bitver latest
+    popd
+}
+
 function download_bits_202206 {
-    bitprefix=$FPGABITS/202206
-    wget_and_link http://doc/cpu_signoff/_downloads/7838d023a633edfce9009ead8f27f808/n300_software_best_config_ddr200t_16M_040a67d4_a1560db8d_202203111641_v3.7.0.bit ${N300_BIT} $bitprefix
-    wget_and_link http://doc/cpu_signoff/_downloads/bcf1baf4265fa729b6d62ab445d55303/ux900_software_best_config_vpu_vcu118_100M_764e6e25f_e63ce562e_202205301028_v2.3.0.bit ${UX900_BIT} $bitprefix
+    local bitprefix=$FPGABITS/202206
+    wget2dir http://doc/cpu_signoff/_downloads/7838d023a633edfce9009ead8f27f808/n300_software_best_config_ddr200t_16M_040a67d4_a1560db8d_202203111641_v3.7.0.bit ${N300_BIT} $bitprefix
+    wget2dir http://doc/cpu_signoff/_downloads/bcf1baf4265fa729b6d62ab445d55303/ux900_software_best_config_vpu_vcu118_100M_764e6e25f_e63ce562e_202205301028_v2.3.0.bit ${UX900_BIT} $bitprefix
 }
 
 download_bits_202206
+
+make_link_latest 202206
